refactor: Move stepHierarchyUtils to steplist/utils.
